pyproject.toml
[project]
name = "ricky-mcp-remote"
version = "0.1.0"
description = "MCP Server for mcp tools I build out"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    "jupyterlab>=4.5.3",
    "mcp[cli]>=1.26.0",
    "numpy>=2.4.1",
    "openai>=2.16.0",
    "openai-agents>=0.7.0",
    "pandas>=3.0.0",
    "pypdf>=6.6.2",
    "python-dotenv>=1.2.1",
]
